requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    make upgrade
#
algoliasearch==1.20.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
amqp==5.2.0
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via openedx-events
beautifulsoup4==4.12.3
    # via -r requirements/base.in
billiard==4.2.0
    # via celery
boto3==1.34.144
    # via django-ses
botocore==1.34.144
    # via
    #   boto3
    #   s3transfer
celery==5.4.0
    # via -r requirements/base.in
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
code-annotations==1.8.0
    # via -r requirements/test.in
coverage[toml]==7.6.0
    # via pytest-cov
ddt==1.7.2
    # via -r requirements/test.in
distro==1.9.0
    # via openai
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
    #   django-choices
    #   django-crum
    #   django-filter
    #   django-model-utils
    #   django-ses
    #   django-solo
    #   django-waffle
    #   djangorestframework
    #   edx-django-utils
    #   openedx-events
django-choices==2.0.0
    # via -r requirements/base.in
django-crum==0.7.9
    # via edx-django-utils
django-filter==23.5
    # via -r requirements/base.in
django-model-utils==4.5.1
    # via -r requirements/base.in
django-object-actions==4.2.0
    # via -r requirements/base.in
django-ses==4.1.0
    # via -r requirements/base.in
django-solo==2.3.0
    # via -r requirements/base.in
django-waffle==4.1.0
    # via edx-django-utils
djangorestframework==3.15.1
    # via -r requirements/base.in
dnspython==2.6.1
    # via pymongo
edx-ccx-keys==1.3.0
    # via openedx-events
edx-django-utils==5.14.2
    # via
    #   -r requirements/base.in
    #   edx-rest-api-client
    #   openedx-events
edx-opaque-keys[django]==2.10.0
    # via
    #   -r requirements/base.in
    #   edx-ccx-keys
    #   openedx-events
edx-rest-api-client==5.7.1
    # via -r requirements/base.in
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
factory-boy==3.3.0
    # via -r requirements/test.in
faker==26.0.0
    # via
    #   -r requirements/test.in
    #   factory-boy
fastavro==1.9.5
    # via openedx-events
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via code-annotations
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.3.7
    # via celery
markupsafe==2.1.5
    # via jinja2
mock==5.1.0
    # via -r requirements/test.in
newrelic==9.12.0
    # via edx-django-utils
openai==1.35.14
    # via -r requirements/base.in
openedx-events==9.11.0
    # via -r requirements/base.in
packaging==24.1
    # via pytest
pbr==6.0.0
    # via stevedore
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.47
    # via click-repl
psutil==6.0.0
    # via edx-django-utils
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via openai
pydantic-core==2.20.1
    # via pydantic
pyjwt==2.8.0
    # via edx-rest-api-client
pymongo==4.8.0
    # via edx-opaque-keys
pynacl==1.5.0
    # via edx-django-utils
pytest==8.2.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements/test.in
pytest-django==4.8.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   faker
python-slugify==8.0.4
    # via code-annotations
pytz==2024.1
    # via
    #   -r requirements/base.in
    #   django
    #   django-ses
pyyaml==6.0.1
    # via
    #   code-annotations
    #   responses
requests==2.32.3
    # via
    #   algoliasearch
    #   edx-rest-api-client
    #   responses
    #   slumber
responses==0.25.3
    # via -r requirements/test.in
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via
    #   edx-ccx-keys
    #   python-dateutil
slumber==0.7.1
    # via edx-rest-api-client
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.5.1
    # via django
stevedore==5.2.0
    # via
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
testfixtures==8.3.0
    # via -r requirements/test.in
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
tqdm==4.66.4
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   asgiref
    #   django-solo
    #   edx-opaque-keys
    #   kombu
    #   openai
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via celery
urllib3==1.26.19
    # via
    #   botocore
    #   requests
    #   responses
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
